Remove the dir the Gomfile is in from GOPATH
None of us can understand why it is there. It makes no sense to us since
the Gomfile is down in a repo, probably somewhere under
  GOPATH/src/github.com/OWNER/REPO/...

In the rarer case there the repo is cloned completely outside of GOPATH
then any sub-packages in the repo can't be accessed properly anyway
(because the 'src/github.com/OWNER/REPO/' path prefix is missing),
so we don't have to consider that case.