Remove the dir the Gomfile is in from GOPATH
None of us can understand why it is there. It makes no sense to us since
the Gomfile is down in a repo, probably somewhere under
  GOPATH/src/github.com/OWNER/REPO/...

In the case that the repo is located outside of GOPATH then any
sub-packages in the repo can't be accessed properly anyway
because the 'src/github.com/OWNER/REPO/' path prefix is missing,
so adding Gompath's dir to GOPATH doesn't break or fix anything.